+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.  
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and follows the [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format.
 
+## [v3.2.1]
+
+### Added
+
+- **Configurable Claim Base URL**
+
+  - Added optional `claimUrl` to `ESPRMBase.configure()` for configuring the base URL used by claim requests.
+
 ## [v3.2.0]
 
 ### Added